Update newrelic-redis to Datastores
New Relic has changed the formats for recording Datastore metrics, and
provided a public API in the agent (as of 3.11.0) that should be used to
record those values.  This commit updates newrelic-redis'
instrumentation to use those API's.

Additionally, this updates the required version of newrelic_rpm
associated with newrelic-redis, as it relies on API's that only were
added in 3.11.0.